Use numbered symbolic constants to represent logic terms, instead of strings
This makes the logic evaluation code much more straightforward and efficient
(no more string-munging). But it also required some changes to how logic is written:

- Transitions can no longer be a condition. We use CreatorBoss, CreatorTrioBoss,
  VirusBoss and MiniBoss symbols as appropriate, instead.
- Numbered item references like Item[0] or Lever[42] are no longer usable as conditions.
  We use specific logic symbols for these items now (for example, Item[0] is now RustedKey)
- In general, all terms in a logic expression must be either a predefined logic symbol,
  the alias of a check in the same room (which resolves to the logic symbol that check
  gives), or a macro name.

The set of logic symbols is defined by the new LogicSymbol enum. Since it's a contiguous set of
numbers, we can use a simple array to represent tables that use them as keys, instead of a dictionary.

There are two special symbols:
- Nil: given by all items that are not significant for logic (like disruptors or lore tablets)
- False: never given by any check, used to denote conditions that can never be satisfied.