Fix random generation failures
When the randomizer tried to pull a location from the pool with only 2 left,
WeigthedSet.RemoveItem would perform an out-of-bounds array access due to
a logic error (which also affected seeds that would seemingly successfully
generate).

This could occur with any combination of settings, but was more likely to
happen without scrap shrines randomized and no skips, probably because those
conditions made it more frequent for there to be only 2 locations left at
any one point during generation.